+++ CHANGELOG.md
@@ -5,6 +5,12 @@ Changelog
 
 # 2.x release
 
+## v2.2.1
+
+- Fix: `compress` flag shouldn't overwrite existing `Accept-Encoding` header.
+- Fix: multiple `import` rules, where `PassThrough` etc. doesn't have a named export when using node <10 and `--exerimental-modules` flag.
+- Other: Better README.
+
 ## v2.2.0
 
 - Enhance: Support all `ArrayBuffer` view types